+++ front/context/src/AccountsContext.tsx
@@ -2,7 +2,6 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-
 import React, { createContext, useEffect, useState } from 'react';
 
 import { InjectedAccountExt } from './types';
@@ -24,7 +23,12 @@ export function AccountsContextProvider(props: Props): React.ReactElement {
   useEffect(() => {
     const getInjected = async (): Promise<void> => {
       if (typeof window !== 'undefined') {
-        const { web3Accounts, web3Enable } = require('@polkadot/extension-dapp');
+        const {
+          web3Accounts,
+          web3Enable,
+          // Can we do `import { } from '@polkadot/extension-dapp'` on the top?
+          // eslint-disable-next-line @typescript-eslint/no-var-requires
+        } = require('@polkadot/extension-dapp');
         await web3Enable('nomidot');
         const [injectedAccounts] = await Promise.all([
           web3Accounts().then((accounts: InjectedAccountExt[]) =>
@@ -44,8 +48,9 @@ export function AccountsContextProvider(props: Props): React.ReactElement {
 
         setInjected(injectedAccounts);
       }
-      getInjected();
     };
+
+    getInjected();
   }, []);
 
   return (