+++ front/context/src/StakingContext.tsx
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { DerivedStakingAccount, DerivedFees } from '@polkadot/api-derive/types';
+import { DerivedFees, DerivedStakingAccount } from '@polkadot/api-derive/types';
 import { Option } from '@polkadot/types';
 import { AccountId } from '@polkadot/types/interfaces';
 import React, { createContext, useContext, useEffect, useState } from 'react';