+++ front/ui-components/src/WalletCard.tsx
@@ -2,8 +2,9 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { Card, Header, Stacked, SubHeader, WithSpaceAround } from './index';
 import React from 'react';
+
+import { Card, Header, Stacked, SubHeader, WithSpaceAround } from './index';
 
 type WalletCardProps = {
   children: React.ReactNode;