+++ front/ui-components/src/stateful/Balance.tsx
@@ -16,10 +16,16 @@ interface BalanceProps
   > {
   address: string;
   detailed?: boolean;
+  orientation?: 'horizontal' | 'vertical';
 }
 
 export function Balance(props: BalanceProps): React.ReactElement {
-  const { address, detailed = false, ...rest } = props;
+  const {
+    address,
+    detailed = false,
+    orientation = 'horizontal',
+    ...rest
+  } = props;
   const { api, isReady } = useContext(ApiContext);
   const [allBalances, setAllBalances] = useState();
   const [allStaking, setAllStaking] = useState();
@@ -33,7 +39,7 @@ export function Balance(props: BalanceProps): React.ReactElement {
         setAllBalances(allBalances);
         setAllStaking(allStaking);
       });
-  
+
       return (): void => balanceSub.unsubscribe();
     }
   }, [api, isReady, address]);
@@ -48,6 +54,7 @@ export function Balance(props: BalanceProps): React.ReactElement {
       allBalances={allBalances}
       allStaking={allStaking}
       detailed={detailed}
+      orientation={orientation}
       handleRedeem={handleRedeem}
       {...rest}
     />