+++ front/ui-components/stories/customDecorators/index.ts
@@ -2,6 +2,8 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import { withApi } from './withApi';
+import { withKeyring } from './withKeyring';
 import { withTheme } from './withTheme';
 
-export { withTheme };
+export { withApi, withKeyring, withTheme };