+++ jest.config.js
@@ -4,5 +4,5 @@
 
 module.exports = {
   preset: 'ts-jest',
-  testPathIgnorePatterns: ['/lib/', '/node_modules/'],
+  testMatch: ['**/*.spec.ts'],
 };